Fix the route autowiring system
It actually seemed the entire system was broken by design!

Routes were never registered and thus never loaded, and
once the CompilerPass responsible for loading was registered,
a circular dependency was revealed.

To ensure this bundle is working and will continue to work,
the tests have been greatly improved and functional tests
were added.

Secondly a resource tracking system was added to ensure,
disabling/enabling a routing import is properly reloaded.
The developer however does need to register the resources
manually.